Refactor: content card 관심사 분리
